requirements.txt
python-dotenv
fastapi
uvicorn
google-genai
python-dotenv
Pillow
